refactor: refactor ui and add router
