+++ src/features/app/index.tsx
@@ -1,0 +1,11 @@
+import { Link } from "react-router-dom";
+
+function App() {
+  return (
+    <div>
+      <Link to="./profile">Profile card</Link>
+    </div>
+  );
+}
+
+export default App;
