+++ src/features/profile-cards/styles.css
@@ -1,0 +1,124 @@
+/*them padding de margin tranh anh huong phan tu cha*/
+.container {
+  background-color: #2eac60;
+  height: 100vh;
+  width: 100vw;
+  padding: 1px;
+}
+
+.cards {
+  margin-left: -50px;
+  display: flex;
+  justify-content: center;
+  margin-top: 200px;
+  flex-wrap: wrap;
+  font-family: sans-serif;
+}
+
+.card {
+  width: 400px;
+  height: 500px;
+  background-color: #242628;
+  display: flex;
+  flex-direction: column;
+  margin-left: 50px;
+  border-radius: 20px;
+}
+
+.card__avatar {
+  border-radius: 50%;
+  width: 160px;
+  background-color: transparent;
+  border: 4px solid #e55d46;
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 20px;
+  cursor: pointer;
+  transition: all 0.2s ease-in-out;
+}
+
+/*set width la % de img ben tron div*/
+img {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
+  background-color: transparent;
+  object-fit: cover;
+}
+
+.card__user-name {
+  font-weight: bold;
+  color: white;
+  font-size: 30px;
+  text-align: center;
+}
+
+.card__user-major {
+  font-weight: bold;
+  color: #e55d46;
+  font-size: 20px;
+  text-align: center;
+  margin-top: -10px;
+}
+
+.card__social-icons {
+  display: flex;
+  justify-content: center;
+  margin-left: -50px;
+  margin-top: 40px;
+}
+
+.card__social-icon {
+  color: white;
+  font-size: 20px;
+  margin-left: 50px;
+}
+
+.card__btn-contact {
+  background-color: transparent;
+  border: 1px #e55d46 solid;
+  border-radius: 5px;
+  padding: auto;
+  color: white;
+  margin: auto;
+  width: 150px;
+  height: 40px;
+  font-size: large;
+  cursor: pointer;
+}
+
+.card__btn-contact:hover {
+  background-color: #e55d46;
+}
+
+.card__avatar:hover {
+  margin-top: 0px;
+  width: 100%;
+  height: 100%;
+  border: none;
+}
+
+/* bo di object fit contain*/
+
+.card__avatar:hover > img {
+  width: 100%;
+  height: 100%;
+  border-radius: 20px;
+  object-fit: cover;
+}
+
+.card__avatar:hover ~ .card__user-name {
+  display: none;
+}
+
+.card__avatar:hover ~ .card__user-major {
+  display: none;
+}
+
+.card__avatar:hover ~ .card__social-icons {
+  display: none;
+}
+
+.card__avatar:hover ~ .card__btn-contact {
+  display: none;
+}
