+++ src/index.css
@@ -1,8 +1,6 @@
 body {
   margin: 0;
   padding: 0;
-  background-color: #2EAC60;
-  font-family: sans-serif;
 }
 
 #root {
