+++ src/pages/AppPage.tsx
@@ -1,0 +1,7 @@
+import App from "../features/app";
+
+function AppPage() {
+  return <App />;
+}
+
+export default AppPage;
