+++ src/pages/ProfileCardPage.tsx
@@ -1,0 +1,7 @@
+import ProfileCard from "../features/profile-cards";
+
+function ProfileCardPage() {
+  return <ProfileCard />;
+}
+
+export default ProfileCardPage;
