+++ src/pages/index.tsx
@@ -1,0 +1,4 @@
+import AppPage from "./AppPage";
+import ProfileCardPage from "./ProfileCardPage";
+
+export { AppPage, ProfileCardPage };
